- moved EnableStackDriver.* to the `autoconfigure` module - created a starter `pom.xml` so that others can build their own collector (and override things like `Credentials` easily) - changed the resolution of `projectId` property
