+++ collector/pom.xml
@@ -7,43 +7,16 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <packaging>jar</packaging>
-
   <artifactId>collector</artifactId>
   <name>Stackdriver Zipkin Collector</name>
-
   <properties>
     <start-class>com.google.cloud.trace.zipkin.StackdriverZipkinCollector</start-class>
   </properties>
-
-  <!--
-  Don't use embedded Tomcat as it conflicts with "netty-tcnative-boringssl-static"
-  See https://github.com/netty/netty-tcnative/issues/151
-  -->
   <dependencies>
     <dependency>
-      <groupId>io.zipkin.java</groupId>
-      <artifactId>zipkin-server</artifactId>
-      <version>${zipkin.version}</version>
-      <exclusions>
-        <exclusion>
-          <groupId>org.springframework.boot</groupId>
-          <artifactId>spring-boot-starter-tomcat</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-jetty</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <artifactId>autoconfigure-storage</artifactId>
-      <version>0.2.1-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>io.netty</groupId>
-      <artifactId>netty-tcnative-boringssl-static</artifactId>
-      <version>1.1.33.Fork23</version>
+      <groupId>com.google.cloud.trace.adapters.zipkin</groupId>
+      <artifactId>starter</artifactId>
+      <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -55,10 +28,9 @@
       <artifactId>zipkin</artifactId>
       <scope>test</scope>
       <type>test-jar</type>
-      <version>${zipkin.version}</version>
     </dependency>
   </dependencies>
-<build>
+  <build>
   <resources>
     <resource>
       <directory>src/main/resources</directory>
